+++ src/serialize/index.ts
@@ -1,3 +1,4 @@
+export * from './default-serializer';
 export * from './serialization';
 export * from './serializer';
 export * from './unsupported-serializer-type-error';