+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 ## Unreleased
 
+## [0.9.0] - 2023-12-10
+
 ### Added
 
 - Support for element traversal via the [Visitor pattern](https://en.wikipedia.org/wiki/Visitor_pattern) ([#14](https://github.com/siren-js/client/issues/14))
@@ -163,6 +165,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
   - `submit()` method for submitting actions; currently, only actions whose `type` is `"application/x-www-form-urlencoded"` are supported.
 - Client `Response` wrapper for parsing Siren responses
 
+[0.9.0]: https://github.com/siren-js/client/compare/v0.8.2...v0.9.0
 [0.8.2]: https://github.com/siren-js/client/compare/v0.8.1...v0.8.2
 [0.8.1]: https://github.com/siren-js/client/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/siren-js/client/compare/v0.7.0...v0.8.0